Switched to SavedModelBuilder.
Instead of using a Saver directly to save the variables, we use the newer SavedModelBuilder. This makes it easier to load the model back in even if you don't know the structure of the original graph. However, the builder doesn't really support saving a model multiple times (e.g. during training each time the model improves). The _load function takes care of this, though, by saving the model into a temporary directory and then moving the files into the actual log directory.
The need for a log file to load a model is gone now. In order to restore the parameters (e.g. uses_dataset), the collections of tensors (e.g. metrics), and individual tensors stored as attributes (e.g. train_op), we convert these to string representations and store them as constants in the graph itself (during _add_savers_and_writers). The appropriate Python variables are recreated during _load.