Changed input_shape param to input_data: includes dtype.
Previously, all inputs were assumed to be tf.float32. To make this more flexible (e.g. when passing in indices to embeddings), one now specifies both the shape and the dtype (as a string like 'int32' or 'float32') for each input. CNN has been updated to construct its placeholders accordingly.