Added layer_input_map to BranchLayer.
This allows one to specify which inputs to a BranchLayer go to which Layer (rather than always having the default of the i'th input going to the i'th layer). This is useful when multiple inputs should go to the same Layer (e.g. if merging some but not all of the branches).