Modified l2, settable learning rate, nicer interrupt for train, CNN->NN, RNN-/>.
BaseNN.train now wraps the main loop in try/except and catches KeyboardInterrupt. This is for aesthetic purposes; now if you stop a model early, it won't display an error message, and it will also still return the metrics.
The learning rate in NN is now a TensorFlow Variable which is passed to the optimizer. This means that if you change the learning rate (using tf.assign), this change will propagate to the optimizer. This is useful, e.g., if you want to decrease the learning rate after the model has converged and then continue training.
The modified version of l2 loss (which is turned on with a flag to NN) divides the l2 loss by the number of trainable parameters in the model. This means that adding more parameters to the model doesn't increase the l2 loss if they're of the same magnitude (on average) as the existing parameters. Note that if using this, you'll probably want to use a much larger value for `l2_lambda` than previously.
The CNN class was renamed to NN to reflect that it isn't CNN specific anymore; this class should be usable for any general neural network.
The RNN class has been removed because it was outdated and should be implementable using the NN class (as long as the labels are already repeated for each timestep before passing them in).