tensor_updates, epoch and early stop metric saved; fixes #5, fixes #7, fixes #8.
The epoch and the best early stop metric value are stored as tensorflow Variables so that they can be saved with the graph and recovered when it's reloaded.
These variables, as well as the learning rate, will often need to be updated. To prevent creating new Assign operations for each update, one Assign op is created for each of these tensors (with a shared placeholder to feed in the new value to assign). The function `update_tensor` makes it easy for a user to pass in a new value and update any of these tensors by name.

A couple of improvements to saving models: the temporary directory used to store the model in before it's moved to the main log directory is created in the log directory itself. This prevents issues that were sometimes (though never earlier...) arising with moving files between disks. If a saved model is being overwritten, only the events files are removed from train/val. This prevents errors where there are other files that shouldn't/can't be removed.