Bidirectional LSTMs, better method of returning last LSTM output.
We take advantage of the fact that, when given sequence_lengths, dynamic_rnn will propagate the last output value (from the last input timestep before padding, as determined by sequence_lengths) forward for the rest of the timesteps in the hidden cell state (the outputs themselves will be set to 0). Taking the last hidden state as the final output saves us the computation of finding and indexing the right timestep to get the final valid output from the array of outputs for all timesteps.